+++ internal/grpcserver/decrypt.go
@@ -10,7 +10,7 @@ import (
 	pb "google.golang.org/protobuf/proto"
 
 	"github.com/nekr0z/muhame/internal/crypt"
-	"github.com/nekr0z/muhame/internal/proto"
+	"github.com/nekr0z/muhame/pkg/proto"
 )
 
 // DecryptInterceptor returns a grpc.UnaryServerInterceptor that decrypts the
